refactor: move onboarding to new packages
Move the onboarding classes into subpackages & rename the OnboardingInteractor to InitialInteractor
