--- app/src/main/java/com/chesire/zwei/view/onboarding/OnboardingInteractor.kt
@@ -1,6 +1,0 @@
-package com.chesire.zwei.view.onboarding
-
-interface OnboardingInteractor {
-    fun completeWelcome()
-    fun completeRequest()
-}
