+++ app/src/main/java/com/chesire/zwei/view/onboarding/initial/InitialInteractor.kt
@@ -1,0 +1,6 @@
+package com.chesire.zwei.view.onboarding.initial
+
+interface InitialInteractor {
+    fun completeWelcome()
+    fun completeRequest()
+}
