+++ app/src/main/java/com/chesire/zwei/view/onboarding/initial/WelcomeFragment.kt
@@ -1,4 +1,4 @@
-package com.chesire.zwei.view.onboarding
+package com.chesire.zwei.view.onboarding.initial
 
 import android.content.Context
 import android.os.Bundle
@@ -9,17 +9,17 @@ import androidx.fragment.app.Fragment
 import com.chesire.zwei.R
 import kotlinx.android.synthetic.main.fragment_welcome.view.buttonNext
 
-class RequestFragment : Fragment() {
-    private lateinit var onboardingInteractor: OnboardingInteractor
+class WelcomeFragment : Fragment() {
+    private lateinit var initialInteractor: InitialInteractor
 
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        return inflater.inflate(R.layout.fragment_request, container, false).apply {
+        return inflater.inflate(R.layout.fragment_welcome, container, false).apply {
             buttonNext.setOnClickListener {
-                onboardingInteractor.completeRequest()
+                initialInteractor.completeWelcome()
             }
         }
     }
@@ -27,13 +27,13 @@ class RequestFragment : Fragment() {
     @Suppress("UnsafeCast")
     override fun onAttach(context: Context?) {
         super.onAttach(context)
-        onboardingInteractor = context as OnboardingInteractor
+        initialInteractor = context as InitialInteractor
     }
 
     companion object {
-        const val tag = "RequestFragment"
-        fun newInstance(): RequestFragment {
-            return RequestFragment().apply {
+        const val tag = "WelcomeFragment"
+        fun newInstance(): WelcomeFragment {
+            return WelcomeFragment().apply {
                 arguments = Bundle()
             }
         }
