+++ app/src/main/java/com/chesire/zwei/view/onboarding/search/SearchInteractor.kt
@@ -1,0 +1,6 @@
+package com.chesire.zwei.view.onboarding.search
+
+interface SearchInteractor {
+    fun completeEnterWorld()
+    fun completeEnterCharacter()
+}
